% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweak.R
\name{tweak_after_pkgdown}
\alias{tweak_after_pkgdown}
\title{Tweak site after \code{pkgdown::build_site()}}
\usage{
tweak_after_pkgdown()
}
\value{
\code{tweak_after_pkgdown()} returns \code{TRUE} invisibly after success.
}
\description{
This consecutively runs \code{\link[=tweak_ref_pages]{tweak_ref_pages()}} and \link{tweak_articles_toc}.
}
