% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweak.R
\name{tweak_articles_toc}
\alias{tweak_articles_toc}
\title{Tweak articles' Table Of Contents}
\usage{
tweak_articles_toc()
}
\value{
Full file paths of (possibly) modified files (invisibly).
}
\description{
This function adds "toc-lev{x}" classes to all elements of TOC, where \code{{x}}
is items "hierarchy" level. For example, if TOC elements are generated with
"h2" and "h3" headers, then it adds "toc-lev1" to "h2" headers and "toc-lev2"
to "h3" headers. \strong{Note} that levels are actually determined by empirical
hack which exploits how 'pkgdown' handles TOC with headers of several orders.
}
\details{
\strong{Notes}:
\itemize{
\item This is often used with \code{padding-left: 20px;} (move elements to the right
by 20 pixels) CSS property of \code{toc-lev2} class in CSS file of package site
(usually 'pkgdown/extra.css'); \code{padding-left: 40px;} for \code{toc-lev3}, etc.
\item To enable having elements of different header order, use "depth: 3" (or
more) property of "toc" element in '_pkgdown.yml'.
}
}
