% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweak.R
\name{tweak_code_blocks}
\alias{tweak_code_blocks}
\title{Tweak Code Blocks}
\usage{
tweak_code_blocks()
}
\value{
Full file paths of (possibly) modified files (invisibly).
}
\description{
Function to tweak code blocks in all relevant \code{pkgdown} pages: index,
reference, and articles. It adds child node \<code class = "r"\> to all
\<pre\> nodes to make highlight.js work. \strong{Note} that this function should
be run with working directory being package root. Also \strong{note} that it will
rewrite pages.
}
