% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweak.R
\name{tweak_ref_pages}
\alias{tweak_ref_pages}
\title{Tweak Reference pages}
\usage{
tweak_ref_pages()
}
\description{
Function to tweak all Reference pages in `pkgdown` site. For now it adds
child node <code class = "r"> to all <pre> nodes to make highlight.js work.
__Note__ that this function should be run with working directory being
package root.
}
