% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweak.R
\name{tweak_after_pkgdown}
\alias{tweak_after_pkgdown}
\title{Tweak site after \code{pkgdown::build_site()}}
\usage{
tweak_after_pkgdown()
}
\value{
\code{tweak_after_pkgdown()} returns \code{TRUE} invisibly after success.
}
\description{
For now this is effectively an alias to \code{tweak_code_blocks()}. This is
because \code{pkgdown} 0.1.5 introduced some important changes:
\itemize{
\item Syntax highlighting process is now the same in index, reference, and
article pages.
\item Tables of contents in sidebars now use \code{bootstrap-toc}, which is good
enough without extra tweaking.
}
}
